+++ tests/Factory/Header/ContentEncodingFactoryTest.php
@@ -7,7 +7,7 @@ use Innmind\Http\{
     Factory\Header\Factory,
     Header\ContentEncoding,
 };
-use Innmind\TimeContinuum\Clock;
+use Innmind\Time\Clock;
 use Innmind\Immutable\Str;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;