+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Requires PHP `8.4`
+- Requires `innmind/filesystem:~9.0`
 - Requires `innmind/time:~1.0`
 
 ## 8.0.0 - 2025-04-20