+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Requires PHP `8.4`
 - Requires `innmind/filesystem:~9.0`
 - Requires `innmind/time:~1.0`
+- `Innmind\Http\TimeContinuum\Format\Http` has been moved to `Innmind\Http\Time\Format\Http`
 
 ## 8.0.0 - 2025-04-20